Introduce team number validation to scout report page
